c库使用优化完成，主要是数学库占空间，缩减到1.5k

+++ src/driver/ms5611.h
@@ -26,8 +26,8 @@ struct ms5611_s
     uint8_t measure_phase;
     uint8_t collect_phase;
     int32_t TEMP;
-    int64_t OFF;
-    int64_t SENS;    
+    int32_t OFF;
+    int32_t SENS;    
 	struct i2c_s* i2c;
 };
